small improvements: work-around cascade, Author exception handling, delete BlogPost updated to GET, updated Author register restriction annotations

+++ src/main/java/be/intecbrussel/blogcentral/controller/AuthorController.java
@@ -20,9 +20,12 @@ import java.util.NoSuchElementException;
 @Controller
 @RequestMapping("/authors")
 public class AuthorController {
+
     private AuthorService authorService;
+
     @Autowired
     private BCryptPasswordEncoder bCryptPasswordEncoder;
+
     private BlogpostService blogpostService;
     private String comingFrom = "";
 
@@ -126,9 +129,32 @@ public class AuthorController {
 
     // delete Author
     @GetMapping("/delete/{id}")
-    public String deleteAuthor(@PathVariable String id) {
+    public String deleteAuthor(@PathVariable String id, Model model) {
 
-        Integer idInt = Integer.parseInt(id);
+        int idInt = convertStringIdToInt(id);
+
+        // help method returns 0 if conversion to nr. didn't work
+        if (idInt == 0) {
+            String noNumber = id + " is not a numeric format";
+            model.addAttribute("error", noNumber);
+            return "error-page";
+        }
+
+        Author authorDB = checkAuthorIdExists(idInt);
+
+        // help method returns null if authorId can not be found
+        if (authorDB == null) {
+            String idNotExist = "no author with id: " + idInt;
+            model.addAttribute("error", idNotExist);
+            return "error-page";
+        }
+
+        List<BlogPost> blogPostsForAuthor =
+                blogpostService.getAllBlogPostFromAuthor(authorDB);
+        for (BlogPost blogPost : blogPostsForAuthor) {
+            blogpostService.deleteBlogPost(blogPost);
+        }
+
         this.authorService.deleteAuthorById(idInt);
         return "redirect:/authors/"; // placeholder
     }
